added several components

+++ app/controllers/blogs_controller.rb
@@ -1,10 +1,14 @@
 class BlogsController < ApplicationController
-  rescue_from ActiveRecord::RecordNotFound, with: :render_not_found_response
-
   # GET /blogs
   def index
     blogs = Blog.all
     render json: blogs
+  end
+
+  # GET /blogs/:id
+  def show
+    blog = Blog.find_by(id: params[:id])
+    render json: blog
   end
 
   # POST /blogs
@@ -13,38 +17,31 @@ class BlogsController < ApplicationController
     render json: blog, status: :created
   end
 
-  # GET /blogs/:id
-  def show
-    blog = find_blog
-    render json: blog
-  end
-
-  # PATCH /blogs/:id
+  # PUT /blogs/:id
   def update
-    blog = find_blog
-    blog.update(blog_params)
-    render json: blog
+    blog = Blog.find_by(id: params[:id])
+    if blog
+       blog.update(blog_params)
+        render json: blog, status: :accepted
+      else
+        render json: { errors: "blog not found" }, status: :not_found
+      end
   end
-
 
   # DELETE /blogs/:id
   def destroy
-    blog = find_blog
-    blog.destroy
-    head :no_content
+    blog = Blog.find_by(id: params[:id])
+    if blog
+      blog.destroy
+      head :no_content
+    else
+      render json: { error: "Blog not found" }, status: :not_found
+    end
   end
 
   private
 
-  def find_blog
-    Blog.find(params[:id])
-  end
-
   def blog_params
-    params.permit(:name, :species)
-  end
-
-  def render_not_found_response
-    render json: { error: "Blog not found" }, status: :not_found
+    params.permit(:title, :description, :image, :author)
   end
 end
